Death is the solution to all problems. No Fusion table - no problem.
Fusion table is killed, no longer needed. Instead use (=update) the copy
in in/nGeigie_map.csv and occasionally push back.

Since this is a temprary hack and the new API will provide realtime
access to all of the data needed to populate the table, in a way
hardcoding is better than relying on 3rd party not-so-database
providers.

NOTE: The CSV file contains multiline comments! Keep those properly
quited, or it WILL break things.